crewrag/requirements.txt
# Core Dependencies
crewai>=0.28.0
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Vector Store
faiss-cpu>=1.7.4

# Document Processing
llama-parse>=0.3.6
pypdf>=3.15.1

# Text Processing
pysbd>=0.3.4

# Utilities
regex>=2023.0.0
agentops>=0.2.0

# Tracking and Logging
colorlog>=6.7.0 